esc-mod/pyproject.toml
[project]
name = "esc-mod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
